testing more videos, forcing type in palette_bgr24_to_bgr565. Will work on refactoring RLE and enforcing max run lengths for safety when decoding and ensuring a proper format

+++ codec/quantizer.py
@@ -13,7 +13,7 @@ def palette_bgr24_to_bgr565(palette24: np.ndarray) -> np.ndarray:
     R5 = R >> 3
    
     bgr565 = (B5 << 11) | (G6 << 5) | R5
-    return bgr565
+    return bgr565.astype(np.uint16)
